AWS feature in good standard. Able to upload
pictures at create. Able to update the create form but picture
still wont update. Will investigate and research more before
continuing. Next step will be verifying with S3 if the image is
really not updating since it could possibly be not displaying
correctly. Will also research a react crop and preview feature for the
pictures.